# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise HRMS API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      summary: Login with email and password
      parameters:
        - name: X-Tenant-ID
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                totpCode:
                  type: string
      responses:
        "201":
          description: Access token, refresh token, session, and safe user profile
...
  /auth/me:
    get:
      summary: Return current tenant auth context
      parameters:
        - name: X-Tenant-ID
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current user context
...